update: tv.php function_tmdb.php to get tv listing by date, pretty limited set update: use internal cache flush method update: limit AJAX Chat images to 50% height 50% width

+++ include/bittorrent.php
@@ -2160,27 +2160,6 @@ function formatQuery($query)
     return $query;
 }
 
-/**
- * @param $dir
- */
-function rrmdir($dir)
-{
-    if (is_dir($dir)) {
-        $objects = scandir($dir);
-        foreach ($objects as $object) {
-            if ($object != '.' && $object != '..') {
-                if (filetype($dir . DIRECTORY_SEPARATOR . $object) == 'dir') {
-                    rrmdir($dir . DIRECTORY_SEPARATOR . $object);
-                } else {
-                    unlink($dir . DIRECTORY_SEPARATOR . $object);
-                }
-            }
-        }
-        reset($objects);
-        rmdir($dir);
-    }
-}
-
 function insert_update_ip()
 {
     global $CURUSER, $cache;
@@ -2203,6 +2182,6 @@ function insert_update_ip()
     }
 }
 
-if (file_exists(ROOT_DIR . 'public' . DIRECTORY_SEPARATOR . 'install')) {
+if (!empty($CURUSER) && $CURUSER['class'] >= UC_STAFF && file_exists(ROOT_DIR . 'public' . DIRECTORY_SEPARATOR . 'install')) {
     $session->set('is-danger', '[h1]This site is vulnerable until you delete the install directory[/h1][p]rm -r ' . ROOT_DIR . 'public' . DIRECTORY_SEPARATOR . 'install' . DIRECTORY_SEPARATOR . '[/p]');
 }
